requirements.txt
Django==1.8
pinax-theme-bootstrap==5.6.0
django-user-accounts==1.3.1
metron==1.3.5
eventlog==0.8.0
django-jsonfield==0.9.13
django-timezones==0.2
pytz==2016.4
django-model-utils==2.5
django-taggit==0.18.0
django-reversion==1.10.1
django-markitup==2.2.2
markdown==2.6.5
django-sitetree==1.5.1
# Pillow==3.2.0
easy-thumbnails==2.3
html5lib==0.9999999
django-bleach==0.3.0
requests==2.10.0
djangorestframework==3.4.7

git+git://github.com/pyconca/symposion.git@2016.6#egg=symposion
git+git://github.com/pinax/pinax-boxes.git@2.1.1

# Deployment
gunicorn==19.6.0
dj-database-url==0.4.1
psycopg2==2.6.1
setproctitle==1.1.10
jinja2==2.8
Fabric==1.11.1

